python/requirements.txt
# OCR engine
pytesseract==0.3.13
Pillow==10.4.0

# PDF handling
pikepdf==9.2.0
pdf2image==1.17.0
pypdf==4.3.0

# Image processing
opencv-python-headless==4.10.0.84
numpy==1.26.4
scipy==1.13.0

# Utilities
python-dotenv==1.0.1
tqdm==4.66.4

# Testing
pytest==8.2.2
pytest-cov==5.0.0
pytest-mock==3.14.0
